Massive restructuring and cachedir support
This is part of work to make it possible to properly install this tool,
in /usr/bin and /usr/lib.

The pristine, read-only version of the modules are now in $libdir
(default: ./modules). boot/ has been moved into modules/.boot, and
tinycore/common into modules/.libtinycore -- so everything except
mkrescue and the README are in $libdir.

At runtime, each referenced module is copied into $cachedir/, default
~/.cache/mkrescue/$module/. The contents of $cachedir are (by default)
what gets copied to the target. $tmpdir is also available for
intermediate files, like ISOs.